Added a csv file for books and medicines that can now be used in the database

+++ src/main/java/com/project/online_book_store/dto/PurchasedBookDTO.java
@@ -17,5 +17,5 @@ public class PurchasedBookDTO {
 
     private int quantity;
 
-    private BigDecimal priceDTO;
+    private double priceDTO;
 }
